chore: Fix eslint.config.js
eslintの設定ファイルを読み込めていなかったので、ダミーファイルを配置し、esmoduleとの中間層を追加しました
