+++ .vscode/settings.json
@@ -1,13 +1,23 @@
 {
-    "eslint.experimental.useFlatConfig": true,
-    "typescript.enablePromptUseWorkspaceTsdk": true,
-    "[typescriptreact]": {
-        "editor.quickSuggestions": {
-            "strings": "on"
-        }
-    },
-    "files.associations": {
-        "*.css": "tailwindcss"
-    },
-    "typescript.tsdk": "node_modules\\typescript\\lib"
+  "[json]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+  },
+  "[jsonc]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+  },
+  "[typescriptreact]": {
+    "editor.quickSuggestions": {
+      "strings": "on"
+    }
+  },
+  "eslint.experimental.useFlatConfig": true,
+  "eslint.options": {
+    "overrideConfigFile": "eslint.config.mjs"
+  },
+  "files.associations": {
+    "*.css": "tailwindcss"
+  },
+  "typescript.enablePromptUseWorkspaceTsdk": true,
+  "typescript.tsdk": "node_modules\\typescript\\lib",
+  "editor.tabSize": 2
 }
