+++ package.json
@@ -1,16 +1,18 @@
 {
   "name": "gourmeijo",
   "version": "0.1.0",
-  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "next lint -c eslint.config.js",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.14.18",
     "next": "14.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
@@ -18,6 +20,7 @@
   "devDependencies": {
     "@eslint/eslintrc": "^2.1.3",
     "@eslint/js": "^8.53.0",
+    "@next/eslint-plugin-next": "^14.0.2",
     "@storybook/addon-essentials": "7.5.3",
     "@storybook/addon-interactions": "7.5.3",
     "@storybook/addon-links": "7.5.3",
@@ -27,9 +30,9 @@
     "@storybook/react": "7.5.3",
     "@storybook/testing-library": "0.2.2",
     "@types/eslint": "^8.44.7",
-    "@types/eslint-config-prettier": "^6.11.3",
     "@types/eslint__eslintrc": "^2.1.1",
     "@types/eslint__js": "^8.42.3",
+    "@types/eslint-config-prettier": "^6.11.3",
     "@types/node": "^20.9.0",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
