+++ postcss.config.js
@@ -1,4 +1,6 @@
-export const plugins = {
-  tailwindcss: {},
-  autoprefixer: {},
+module.exports = {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
 };
